feat: add support for recursively scanning directories
This means you can now eg navigate to `~/Programming` and run `dura watch` to watch the whole directory recursively. The program will scan each child dir until either a repo is found or the max depth is hit.

Includes, excludes and the max depth can be manually tweaked in the `config.json` file.