Fix test stability & speed (#77)
* Fix test stability & speed

I noticed that startup tests fail on very slow machines, like the build
hosts. I also hate that there's a 6 second `thread::sleep`. This PR fixes both of
those problems. Tests run in <1s.

To do this, I created `Daemon`, that sets up a mpsc (technically an spsc)
channel that sends each line as a message. This lets tests do a
`read_line` to wait until output is printed. It's a great way to
communicate between parent and child processes.

* Extracted constant

* yeah, but, like, run tests first