Optimization: Check file timestamps (#127)
The current version performs a `git commit` on every repository every 5
seconds. If nothing changed, it simply has no effect. This strategy is
eating up sometimes 10-20% of my CPU (on a poor dev machine), due to the
massive amount of hashing needed to commit so many repositories.

This change creates an optimization by tracking file changed timestamps
and comparing them against last commit timestamps. To do this, we have
to traverse all files in each repository, which `git commit` is doing
anyway. The big difference is that we only need to traverse to the inode
(directory) for the file timestamp entries, so it should be
significantly faster.

Considered design: It has been suggested to use the `notify` crate to be
notified of file changes. The trouble here was a global limit on file
descriptors that can be watched, so you can't watch all files. I
considered some machine learning appraches to guessing which files
should be watched, but abandoned the approach because every strategy I
could think of inolved a fallback to eactly what I'm doing here.

Considered design: Use SQLite to remember timestamps of every file. This
was needlessly complex because I only need to remember one timestamp -
the most recent one. And more important, the last dura commit already
stores it, so SQLite wasn't adding any value.